requirements.txt
flask
flask-cors
python-dotenv
pdfplumber
requests
